+++ test/test-check-kit.js
@@ -1,9 +1,16 @@
-import check, { loadPackageJson } from '../dist/index';
+import chai, { expect } from 'chai';
+import chaiAsPromised from 'chai-as-promised';
+import check, { loadPackageJson } from '../src/index.js';
 import fs from 'fs-extra';
 import http from 'http';
 import path from 'path';
 import snooplogg from 'snooplogg';
 import tmp from 'tmp';
+import { fileURLToPath } from 'url';
+
+chai.use(chaiAsPromised);
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const { log } = snooplogg('test:check-kit');
 const { highlight } = snooplogg.styles;
@@ -11,6 +18,10 @@ const { highlight } = snooplogg.styles;
 process.env.TEST_META_DIR = tmp.dirSync({ prefix: 'check-kit-test' }).name;
 
 describe('check-kit', function () {
+	beforeEach(() => {
+		process.env.FORCE_UPDATE_NOTIFIER = 1;
+	});
+
 	afterEach(async () => {
 		await fs.remove(process.env.TEST_META_DIR);
 	});
@@ -69,6 +80,7 @@ describe('check-kit', function () {
 	describe('Environment check skip', () => {
 		it('should skip check if NO_UPDATE_NOTIFIER is set', async () => {
 			try {
+				delete process.env.FORCE_UPDATE_NOTIFIER;
 				process.env.NO_UPDATE_NOTIFIER = 1;
 				expect(await check()).to.deep.equal({});
 			} finally {
@@ -78,6 +90,7 @@ describe('check-kit', function () {
 
 		it('should skip check if NODE_ENV is set to test', async () => {
 			try {
+				delete process.env.FORCE_UPDATE_NOTIFIER;
 				process.env.NODE_ENV = 'test';
 				expect(await check()).to.deep.equal({});
 			} finally {
@@ -126,7 +139,9 @@ describe('check-kit', function () {
 			expect(result.updateAvailable).to.equal(true);
 		});
 
-		it('should check a valid package.json with scoped name', async () => {
+		it('should check a valid package.json with scoped name', async function () {
+			this.timeout(5000);
+
 			const result = await check({
 				pkg: path.resolve(__dirname, 'fixtures/good-scoped/package.json')
 			});